requirementes.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dataclasses==0.8
distlib==0.3.2
Django==3.2.11
django-cors-headers==3.10.1
djangorestframework==3.13.1
filelock==3.0.12
flake8==3.9.2
idna==3.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
orjson==3.6.1
pathspec==0.9.0
platformdirs==2.2.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.11.2
pylint==2.9.3
PyQt3D==5.15.5
PyQt3D-Qt5==5.15.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtChart-Qt5==5.15.2
PyQtDataVisualization==5.15.5
PyQtDataVisualization-Qt5==5.15.2
PyQtNetworkAuth==5.15.5
PyQtNetworkAuth-Qt5==5.15.2
PyQtPurchasing==5.15.5
PyQtPurchasing-Qt5==5.15.2
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
pytz==2021.3
QScintilla==2.13.1
rapidjson==1.0.0
regex==2021.7.6
requests==2.26.0
serpy==0.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.1.0
typed-ast==1.4.3
typing_extensions==4.0.1
urllib3==1.26.6
virtualenv==20.7.2
wrapt==1.12.1
zipp==3.6.0
